move request helper functions out of NodeExecutionFunctions
