+++ packages/core/src/execution-engine/node-execution-context/execute-single-context.ts
@@ -13,9 +13,6 @@ import type {
 } from 'n8n-workflow';
 import { ApplicationError, createDeferredPromise, NodeConnectionType } from 'n8n-workflow';
 
-// eslint-disable-next-line import/no-cycle
-import { getRequestHelperFunctions } from '@/node-execute-functions';
-
 import { BaseExecuteContext } from './base-execute-context';
 import {
 	assertBinaryData,
@@ -23,6 +20,7 @@ import {
 	getBinaryDataBuffer,
 	getBinaryHelperFunctions,
 } from './utils/binary-helper-functions';
+import { getRequestHelperFunctions } from './utils/request-helper-functions';
 import { returnJsonArray } from './utils/return-json-array';
 
 export class ExecuteSingleContext extends BaseExecuteContext implements IExecuteSingleFunctions {
