+++ packages/core/src/execution-engine/node-execution-context/hook-context.ts
@@ -11,10 +11,8 @@ import type {
 } from 'n8n-workflow';
 import { ApplicationError } from 'n8n-workflow';
 
-// eslint-disable-next-line import/no-cycle
-import { getRequestHelperFunctions } from '@/node-execute-functions';
-
 import { NodeExecutionContext } from './node-execution-context';
+import { getRequestHelperFunctions } from './utils/request-helper-functions';
 import { getNodeWebhookUrl, getWebhookDescription } from './utils/webhook-helper-functions';
 
 export class HookContext extends NodeExecutionContext implements IHookFunctions {
